refactor(bevy_cef): extract is_pixel_transparent into shared alpha module
